+++ src/twg2/jbcm/Op.java
@@ -1,26 +1,26 @@
 package twg2.jbcm;
 
-import twg2.jbcm.modify.OpcodeChangeCpIndex;
-import twg2.jbcm.modify.OpcodeChangeOffset;
+import twg2.jbcm.modify.CpIndexChanger;
+import twg2.jbcm.modify.CodeOffsetChanger;
 
 /**
  * @author TeamworkGuy2
  * @since 2014-4-20
  */
 public final class Op {
-	private OpcodeChangeCpIndex cpIndex;
-	private OpcodeChangeOffset codeOffset;
+	private CpIndexChanger cpIndex;
+	private CodeOffsetChanger codeOffset;
 
 	private Op() {}
 
 
-	public final Op add(OpcodeChangeCpIndex cpIndex) {
+	public final Op add(CpIndexChanger cpIndex) {
 		this.cpIndex = cpIndex;
 		return this;
 	}
 
 
-	public final Op add(OpcodeChangeOffset codeOffset) {
+	public final Op add(CodeOffsetChanger codeOffset) {
 		this.codeOffset = codeOffset;
 		return this;
 	}
@@ -32,13 +32,13 @@ public final class Op {
 	}
 
 
-	public static final Op of(OpcodeChangeCpIndex cpIndex) {
+	public static final Op of(CpIndexChanger cpIndex) {
 		Op op = new Op();
 		return op.add(cpIndex);
 	}
 
 
-	public static final Op of(OpcodeChangeOffset codeOffset) {
+	public static final Op of(CodeOffsetChanger codeOffset) {
 		Op op = new Op();
 		return op.add(codeOffset);
 	}