+++ src/twg2/jbcm/modify/IndexChanger.java
@@ -4,7 +4,7 @@ package twg2.jbcm.modify;
  * @author TeamworkGuy2
  * @since 2014-4-20
  */
-public interface IndexChange {
+public interface IndexChanger {
 
 	public void changeCpIndex(short oldIndex, short newIndex);