+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project does its best to adhere to [Semantic Versioning](http://semver.org/
 
 
 --------
-###[0.1.0](N/A) - 2017-01-11
+### [0.2.0](N/A) - 2017-08-13
+#### Changed
+* Cleaned up unnecessary twg2.jbcm.opcode package classes
+* Renamed twg2.jbcm.modify.IndexChange -> IndexChanger
+* Renamed twg2.jbcm.modify.OpcodeChangeCpIndex -> CpIndexChanger
+* Renamed twg2.jbcm.modify.OpcodeChangeOffset -> CodeOffsetChanger
+
+
+--------
+### [0.1.0](https://github.com/TeamworkGuy2/ClassLoading/commit/dc1eda0b07c1d9746b8a63affff8d8b7b0d2103f) - 2017-01-11
 #### Added
 * Initial versioning of existing code, including classfile loading/parsing, dynamic classfile modification, and runtime class loading.